refactor: don't use re-exports (#568)
If we depend on the crate directly we should not use re-exports, but
rather import the necessary types and functions directly.

This is because re-exports can lead to confusion and make it harder to
understand the dependencies of a crate.

For example, this highlights that we only depend on `reth_evm` for the
`TransactionEnv` trait in `tempo_revm`.