fix(consensus): fill holes when forwarding finalized blocks (#1173)
This patch has the application backfill holes on the execution layer if
it attempts to finalize a block (received from the consensus layer) that
is not immediately on top of the highest finalized height of the
execution layer.

This patch is necessary because
`marshal::Mailbox::get_info(Identifier::Latest)` was misunderstood: it
did not return the latest finalized block that the marshal actor has
forwarded and was acknowledged by the application, but the finalized tip
of the overall network.

This led to crash-loops of the node because it was attempting to
backfill blocks that the marshal actor did not yet have available (and
that should be forwarded by different means).